// gdpicture-client/src/app/models/DocuViewareConfig.ts
export interface DocuViewareConfig {
    SessionId: string
    ControlId: string
    AllowPrint: boolean
    EnablePrintButton: boolean
    AllowUpload: boolean
    EnableFileUploadButton: boolean
    CollapsedSnapIn: boolean
    ShowAnnotationsSnapIn: boolean
    EnableRotateButtons: boolean
    EnableZoomButtons: boolean
    EnablePageViewButtons: boolean
    EnableMultipleThumbnailSelection: boolean
    EnableMouseModeButtons: boolean
    EnableFormFieldsEdition: boolean
    EnableTwainAcquisitionButton: boolean
  }
  